+++ src/test/java/domain/menu/MenuTest.java
@@ -1,7 +1,5 @@
 package domain.menu;
 
-import domain.type.Category;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
