+++ src/test/java/domain/menu/MenusTest.java
@@ -1,7 +1,5 @@
 package domain.menu;
 
-import domain.type.Category;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
