+++ src/test/java/domain/table/TableTest.java
@@ -1,7 +1,7 @@
 package domain.table;
 
 import domain.menu.Menu;
-import domain.type.Category;
+import domain.menu.Category;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
