+++ src/test/java/domain/table/order/OrdersTest.java
@@ -1,7 +1,7 @@
 package domain.table.order;
 
 import domain.menu.Menu;
-import domain.type.Category;
+import domain.menu.Category;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
